modified

+++ src/main/resources/schema.sql
@@ -1,4 +1,4 @@
 create table price_data (
 instrument varchar(10),
-price decimal
+price double
 );
